Refactor: 카카오맵 분리 - 물품상세, 물품등록 Related to: #6
